+++ www/src/components/ThemeEditor/Form.tsx
@@ -23,36 +23,3 @@
  SOFTWARE.
 
  */
-
-import React from 'react'
-import styled from 'styled-components'
-import { Code, Flex, FlexItem } from '@looker/components'
-
-interface PropsProps {
-  of: string
-  interfaceName?: string
-}
-
-const Props = ({ of, interfaceName = `${of}` }: PropsProps) => {
-  return (
-    <Layout>
-      <FlexItem>
-        Interface of <PropsCode>{interfaceName}</PropsCode>
-      </FlexItem>
-      <FlexItem ml="auto">🏗 Coming Soon</FlexItem>
-    </Layout>
-  )
-}
-
-const PropsCode = styled(Code).attrs({ fontSize: 'small' })`
-  color: ${(props) => props.theme.colors.key};
-`
-
-const Layout = styled(Flex).attrs({ mb: 'large', mt: 'small', py: 'small' })`
-  border-top: 1px solid ${(props) => props.theme.colors.ui2};
-  border-bottom: 1px solid ${(props) => props.theme.colors.ui2};
-  color: ${(props) => props.theme.colors.text2};
-  font-size: ${(props) => props.theme.fontSizes.small};
-`
-
-export default Props