+++ www/src/components/ThemeEditor/ThemeEditor.tsx
@@ -24,39 +24,50 @@
 
  */
 
+import { IconButton, DialogManager } from '@looker/components'
 import React, { FC } from 'react'
 import styled from 'styled-components'
-import { Link, ListItem } from '@looker/components'
+import {
+  CoreColors,
+  FontFamilyChoices,
+  IntentColors,
+} from '@looker/design-tokens'
+import {
+  ThemeEditorContent,
+  ThemeEditorContentProps,
+} from './ThemeEditorContent'
 
-const Resource: FC<{ url: string }> = ({ children, url }) => (
-  <Link href={url} target="_blank" rel="noopener noreferrer">
-    <Style>{children}</Style>
-  </Link>
+interface ThemeEditableProps {
+  colors?: Partial<CoreColors & IntentColors>
+  fontFamilies?: Partial<FontFamilyChoices>
+}
+
+export interface ThemeEditorProps extends ThemeEditorContentProps {
+  hasCustomTheme: boolean
+  updateTheme: (customTheme?: ThemeEditableProps) => void
+}
+
+export const ThemeEditor: FC<ThemeEditorProps> = ({
+  hasCustomTheme,
+  updateTheme,
+}) => (
+  <DialogManager
+    content={<ThemeEditorContent updateTheme={updateTheme} />}
+    width="90%"
+    maxWidth="90%"
+  >
+    <IconButtonActivated
+      icon="Beaker"
+      label="Customize Theme"
+      tooltipDisabled
+      activated={hasCustomTheme && 'active'}
+      color={hasCustomTheme ? 'key' : 'neutral'}
+      size="small"
+      round
+    />
+  </DialogManager>
 )
 
-const Style = styled(ListItem).attrs({
-  fontSize: 'small',
-  py: 'xsmall',
-})`
-  align-items: center;
-  color: ${(props) => props.theme.colors.text2};
-  display: flex;
-
-  svg {
-    margin-right: ${(props) => props.theme.space.xsmall};
-    width: 1.5rem;
-    transition: ${(props) =>
-      `transform ${props.theme.transitions.durationModerate} ${props.theme.easings.easeOut}`};
-  }
-
-  &:hover {
-    color: ${(props) => props.theme.colors.text3};
-    text-decoration: none;
-
-    svg {
-      transform: scale(1.2);
-    }
-  }
+const IconButtonActivated = styled(IconButton)<{ activated?: boolean }>`
+  background: ${({ activated, theme }) => activated && theme.colors.keyAccent};
 `
-
-export default Resource