+++ CHANGELOG.md
@@ -37,8 +37,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow `Icon` colors to be modified when used inside a `MenuItem` detail prop.
 - Miscellaneous Storybook fixes to silence console warnings
 
-###
+### Remove
 
+- `Sidebar*` components removed
 - `Tooltip` & `Popover` no longer support (optional) arrow indicator
 
 ### Removed